Use common email template for errors and issue reporting
