+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/error/error.component.ts
@@ -1,6 +1,7 @@
 import { MDC_DIALOG_DATA, MdcDialogRef } from '@angular-mdc/web';
 import { Component, Inject } from '@angular/core';
 import { environment } from 'src/environments/environment';
+import { issuesEmailTemplate } from 'xforge-common/utils';
 
 export interface ErrorAlert {
   message: string;
@@ -19,10 +20,6 @@ export class ErrorComponent {
   constructor(public dialogRef: MdcDialogRef<ErrorComponent>, @Inject(MDC_DIALOG_DATA) public data: ErrorAlert) {}
 
   get issueMailTo(): string {
-    return encodeURI(
-      `mailto:${environment.issueEmail}?subject=${environment.siteName} issue&body=\n\nError id: ${
-        this.data.eventId
-      } (included so we can provide better support)`
-    );
+    return issuesEmailTemplate(this.data.eventId);
   }
 }
